RDKB-6550, RDKB-5986: Log messages related to ath1 going down
Reason for change: In field, we observe that many devices print ath1 is off even when they are Up.
It was observed that just beofre SSID status check, dmcli crashes. As the Self heal depends on the return status of
previous command it interpreted crash as ath1 off condition.

Modified code so that Self heal will depend on strings values explicitly rather on return code.

Test Procedure:

Verified the fix on target by invoiking the script manually.
Prints that were observed in Self heal log file when status is not got due to a process crash is:
[RKDB_PLATFORM_ERROR] : dmcli crashed or something went wrong while checking 2.4G status.
[RKDB_PLATFORM_ERROR] : dmcli crashed or something went wrong while checking 2.4G status.

Log message when the WiFi status is Down:
[RKDB_PLATFORM_ERROR] : 2.4G private SSID (ath0) is off.
[RKDB_PLATFORM_ERROR] : 5G private SSID (ath1) is off.

Change-Id: I5135a94f011cd0cf13ac18dce80876cb23ac05a4
Risks: None
Signed-off-by: achelakkode<anoopnair_c@infosys.com>