RDKB-8065:Marker needed to track devices which is stuck at erouter rebinding
Reason for change:
In field devices, we have observed that certain devices are not able to go out through erouter0 interface. In those units even normal ping were not working.
When these devices are in this state, it is observed that a DHCP negotiation was in place and the interface never got it's IP updated.
To track the percentage of this occurrence, a telemetry marker is needed.

Test Procedure:
1) Verified in unit by manually runnign the script
2) Injected error values to the script and verified that when ping fails and following conditions meet:
   a) erouter0 doesn't have ip
   b) DHCP status is "Rebinding"
we print a log line:
[2016-10-05:16:54:21:760000] EROUTER_DHCP_STATUS:Rebinding

Change-Id: I03807b3a0a4720dc4653f890e10263dddd06ce0e
Risks: None
Signed-off-by: achelakkode<anoopnair_c@infosys.com>