RDKB-8597: Make downstream CPU % parsing independent of the position
Reason for change:
Through RDKB-8578, we tried to resolve an issue where unit didn't take any corrective action even though downstream manager was consuming more than 25% CPU.
Command to parse the CPU utilization is using the position of % now which we need to modify.

Test Procedure:
As the issue is not reproducible, we have manually picked some processes with "N" and without "N" and parsed the output for verification.

Change-Id: I700ff26530858435c03d54ce03a3300d29f8906d
Risks: None
Signed-off-by: achelakkode<anoopnair_c@infosys.com>