RDKB-9430: Lot of devices reporting "No IPv6 Gateway Address detected”
Reason for change: Some units were reporting "No Ipv6 gateway error" eventhough erouter is in dual stack mode.
Issue was with the sysevent which is checked to get the erouter0 ip. Somehow this event is set to be wrong as "::" from dhcp plugin.
Fix is to get the ipv6 address directly from the interface.

Test Procedure:
Verified the fix on an unit from field which was giving no ipv6 gateway address error.

To verify the fix, check in SelfHeal log file in /rdklogs/logs directory that we dont see a print like below every 1 hour:
No IPv6 Gateway Address detected
Risks: None
Signed-off-by: achelakkode<anoopnair_c@infosys.com>

Change-Id: Ib9bdc61838b7b45b4ddc7c197b3596ed10cc026e