RDKB-13705 : [Field] brlan0 is not up.
Reason for change:
From the field box, we observed brlan0 and brlan1 interface not coming up because of 02_multinet event registration not done. Currently we are unable to find the root cause for this case. As of now we put selfheal to re-register 02_multinet events.
Test Procedure:
We observed this issue on field device only. so we can't verify this in local environment. But ensure as usual functionalities will not broke.
Risks: None
Signed-off-by: Lakshminarayanan Shenbagaraj Lakshminarayanan_Shenbagaraj@comcast.com

Change-Id: Ifc567998927e96f7d11579f1d83774d8967df790