RDKB-14653 : CPE is returning param not exist leading to activation failures.
Reason for change:
I looked at most of the devices. Below are 2 reasons why selfheal is stuck.
We do SSH to ATOM from selfheal, in most of the devices SSH is stuck. Usually when peer is not reachable SSH exits with “No Route to Host” error. But in this particular case it is not existing. In dropbear server we have AUTH_TIMEOUT set to 300 secs, in this scenario it is not honoring this TIMEOUT. I tried with idle time out and connection timeout options to ssh, it is not helping. I will spend some time and see if we can timeout SSH in this scenario. But it is just a fix for selfheal stuck issue. so we need to do ping-peer test before ssh test to avoid selfheal stuck.
Test Procedure:
1. Query below WiFi parameter and it should give proper value.
dmcli eRT getv Device.WiFi.SSID.1.Enable
2. After 10mins do hung(stop) WiFi process based on below command,
kill -19 <pid of WiFi>
3. After 15mins from bootup(first task_health_monitor script scheduled).
Ensure in selfheal logs, ping_peer or peer_ping operation should finish before ATOM stuck or WiFi stuck case.
Risks: None
Signed-off-by: Lakshminarayanan Shenbagaraj Lakshminarayanan_Shenbagaraj@comcast.com

Change-Id: I9b5215746acaefb18c938888a535f6ff40005933