MESH-448: Strip mesh pods from traffic data reporting
Reason for change: The cloud team is observing incorrect traffic
data counting because they are counting both STB and NOC results.
So we are going to strip out the mesh pod traffic data from the
router results.

Test Procedure: Verified that the pod mac addresses do not show up
in the reported traffic data.

Risks: Low

Change-Id: I93fbccdc2b6094bc2d36326759d26d4bf4893029
Signed-off-by: scott_angel@cable.comcast.com