TCXB6-4352: SNMP not able to fetch rdkbRgDot11WepPassPhrase
Reason for change: The actual root cause for the problem is,
the WIFI Agent not completing the initialization. In that timeframe
no parameter from WIFI Agent could able to fetch. Currently we don’t have any
self heal mechanism to restart the WIFI Agent. That is causing the device unpollable
for the WIFI parameters.
Now added a check to verify the /tmp/wifi_initialized file present or not, if the
file not present it will restart the CcspWifiSsp

Test Procedure: Made a build checked the Selfheal hasn’t any break
	Removed the /tmp/wifi_initialized and verified the CcspWifiSsp getting
	restarted.
Risks: None
Signed-off-by: Ravi <ravi_veldanda@comcast.com>

Change-Id: I8779ba31357bd5c169beaf9e4396721827baf623