RDKBDEV-405: Some improvement on TestAndDiagnostic module
Reason for change:

fix misuse of "\0"

"\0" is a pointer to an empty string (ie effectively the same as ""),
and is NOT the same as a nul charactor (ie a char with value 0).

Usage of "\0" is almost always a bug. If a pointer to an empty string
is needed then use "", if a nul character is needed then use 0 or
'\0'.

Signed-off-by: Andre McCurdy <armccurdy@gmail.com>

avoid using uninitialised data if traceroute command fails in tracert_start()

Signed-off-by: Andre McCurdy <armccurdy@gmail.com>
(cherry picked from commit 7c781d877aba0fcb316eb757f09a433281089c4c)

declarations for g_GetParamValueString etc in header files should be extern

Variables should be instantiated once in a .c file and declared as
extern in the associated header.

Test Procedure: Build and Sanity
Risks: None

Signed-off-by: Andre McCurdy <armccurdy@gmail.com>
(cherry picked from commit d985b679845383b6111376e70a7f669b2011d5ad)
Change-Id: I31938c5e8e63c2aeccfd6bff0d344adcbf12f0ac