task_health_monitor.sh: Mv1 specific hotspot selfheal support ( RDKB-4987 )
[Description]
Tunnel is getting destroyed after some time even though the endpoint is reachable
Reason was selfheal sending wrong message to recover the bridge structure even though it was correct.
Selfheal validates the bridge structure based on RDKB, but here bridge structure is as per the MV1
Corrected the selfheal for MNG units.

[Test Description]:
-Enable the Hotspot
-Make sure tunnel is created
-Bridge structure is correct
-Wait for selfheal to trigger.
-After the selfheal is triggered, it should not send error message and try to recover the tunnel

Hotspot should be working as expected.

Signed-off-by: Karthik Shanbhag(kshanbhag@libertyglobal.com