RDKCOM-2129 RDKBDEV-410: ensure busybox pidof command is used
Reason for change:
ensure busybox pidof command is used(and not the procps version)
The busybox and procps versions of pidof behave differently when used
to determine the PID of a script (the procps version requires the -x
command line option, which the busybox version neither requires or
supports).
The calls to pidof within the test-and-diagnostic code intended to
run on the target (ie everything except the scripts etc under
source-pc) assume the busybox syntax and may therefore fail to run
correctly if the procps version of pidof is found on the target (e.g.
in a debug or development build).
To ensure consistent behaviour regardless of whether the procps
version of pidof is installed, always run the busybox version
explicitly (ie run "busybox pidof" instead of "pidof").
Note that some of the calls to pidof are made as part of RPC calls
to the second CPU in Puma6 / Puma7 based systems. These have also
been updated to run the busybox version of pidof (which is assumed to
be used on all CPUs).
Test Procedure: Build and Basic sanity
Risks: None.
Signed-off-by: Andre McCurdy <armccurdy@gmail.com>

(cherry picked from commit c14a131928e728629d894e6e5b9042f220e9b393)
Change-Id: Ia961f5ebbe43241a7eff9a963fe33caf301e9bd1