RDKB-38092 : Notify Component to store notification list for recovery
Reason for change: When notify_comp is not storing notification list in some file, it is in local memory. So, if it crashes and restart , it will not have list of registered parameters, implement a new solution, to store it in some file and read it from that file whenever notify_comp comes up.

Test Procedure: Kill notify_comp process. Once self heal restarts it, the PA notifications should happen as before for a respective parameter change. The NOTIFYLog.txt.0 can be checked to see if the file has been reloaded.

Risks: None

Signed-off-by: kavya_chowdahallisuresh@comcast.com
(cherry picked from commit 8dbbc0ef685aff10f6bc5aad9cd526b44dc63596)
Change-Id: Icc940fd79f07627e20c2a2df90f1cfc20803a8c6