task_health_monitor.sh: temp workaround for process cache containing stale PID when CcspPandMSsp is restarted ( OFW-2394 F3896SI-748 )
Fix CcspPandMSsp process is starting two times after we kill it in dualstack mode.

[Description]:
When CcspPandMSsp process is killed, it should be reset. When the
reset happens the process cache is not updated with latest pid with
get_pid. So the script is triggering the reset another time. As a
result it is showing the CcspPandMSsp two times.

[Test Description]:
With the current changes verified the CcspPandMSsp process is showing
one time after kill it in dualstack mode.

Signed-off-by: Narayana DVVS <ndvvs.contractor@libertyglobal.com>