task_health_monitor.sh: drop obsolete support for ssam-agent ( OFW-2571 )
Issue Description:
ssam_agent should be removed from selfheal check. This code is moved to PandM.

Fix Description:
The ssam_agent code is already moved to PandM but still the ssam_agent pid is being checked while the selfheal check. So we have removed the part of the code for
checking the ssam_agent pid.