task_health_monitor.sh: enable WiFi agent health check ( OFW-3060 )
Check if WiFi agent is not hanging in the sealfheal script

[Issue]
In case WiFi agent is stack, sealfheal is not restarting the process.

[Solution]
Add monitoring of wifi agent in selfheal script

[Testing Performed]
Excute the following cammand and check if the WiFi agent is restarted or
not :
kill -s STOP PID_WIFI_AGENT

Next run of selfheal should trigger a restart in the wifi agent.

Upstream Candidate: No. LGI specific fix.

Signed-off-by: Oussema ben abdallah(obenabdallah@libertyglobal.com)