task_health_monitor.sh: extend dibbler-client recovery mechanism for all platforms ( OFW-4242 F5685SI-1126 )
fix Dibbler process is not automatically recovered by the selfheal.

[Issue Description]
Dibbler process is not automatically recovered by the selfheal.

[Root Cause]
For the MV3 boxtype, the code to restart dibbler client during a
crash or killing of process is not covered.

[Solution]
Added dibbler client restart support for MV3 and tested the scenario.
Now, the dibbler client restarts successfully.

Signed-off-by: mjothimani.ext <mjothimani.ext@libertyglobal.com>