task_health_monitor.sh: fix the typo in the if command to check if wanmanager process exists or not ( OFW-5309 F3896SI-2343 )
[Issue]
Wanmanager process is not getting restarted automatically if the old process got killed.

[Solution]
In the selfheal script , there was already a check for wanmanager process and restart if not running.
But the changes were under MV3 flag.
The change in now made generic for MV2+ and MV3 platforms.

[Testing]
Killed the old wanmanager process and ensure that selfheal restarts wanmanager.
Ensured that multiple dibbler client and server process are not running.

Signed-off-by: Keerthana V J(kvijayakumarjalajaku.contractor@libertyglobal.com)