add quotes to ping/traceroute host when value is used ( OFW-105 )
Adding quotes should be done at the time the hostname is used, not as
part of validating it. Cleaning this up allows various workarounds to
strip quotes to be dropped.

Also move host option to the normal location at the end of ping and
traceroute command lines (ie make it the last option, not the first).

Do not upstream: unexplained regression on certain Comcast platforms?

Signed-off-by: anil shetty <adshetty@libertyglobal.com>
Signed-off-by: Andre McCurdy <armccurdy@gmail.com>