resource_monitor.sh: reboot with reason ATOM_HIGH_LOADAVG after 4 retries instead of 1 ( OFW-5683 )
[Issue]
Currently the device goes for reboot with reason ATOM_HIGH_LOADAVG as soon as the load avg is 5 for 1 time.
No retry iterations handled. Now go for reboot after 4 failures.

[Description]
This is not the fix, but wait some more time before going for a reboot.
This change helps to debug the issue on why ATOM side load is high.

[Testing]
1. Manually introduce the load on ATOM with below command:
for i in 1 2 3 4; do while : ; do : ; done & done
3. Look for the ATOM side loadavg value using command
cat /proc/loadavg
2. Verify the reboot happens only after 4 fail count from ARM side self heal.

Upstream Candidate - No

Signed-off-by: Sidda(s.siddalingesh@infosys.com)